requirements.txt
streamlit
requests
python-dotenv
flask
flask-cors
